api/requirements.txt
numpy==1.19.5
ffmpeg-python==0.2.0
fastapi[all]